restructure code to keep all functions together in the samee js source add function jsdocs to spec
