+++ .gitignore
@@ -19,3 +19,5 @@ build
 lib/
 
 doc/grammar.g4
+doc/functions.md
+doc/functions.adoc
