+++ .vscode/settings.json
@@ -12,6 +12,7 @@
     "cssclass",
     "dcba",
     "docdate",
+    "eomonth",
     "EXPREF",
     "firstbar",
     "foobarbaz",
