+++ src/Lexer.js
