+++ src/TreeInterpreter.js
