+++ src/json-formula.js
@@ -9,7 +9,7 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-import Formula from './jmespath/jmespath.js';
+import Formula from './interpreter.js';
 
 /**
  * Returns an instance of JSON JsonFormula Expression that can be executed later on with
