+++ src/stringToNumber.js
