+++ test/extensions.spec.js
@@ -10,7 +10,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 import JsonFormula from '../src/json-formula.js';
-import stringToNumber from '../src/jmespath/stringToNumber.js';
+import stringToNumber from '../src/stringToNumber.js';
 import createForm from '../tutorial/Form.js';
 
 test('if executes correct branch', () => {
