+++ test/jmespath/jmespath.spec.js
@@ -28,7 +28,7 @@ governing permissions and limitations under the License.
 
 import JsonFormula from '../../src/json-formula.js';
 import createForm from '../../tutorial/Form.js';
-import stringToNumber from '../../src/jmespath/stringToNumber.js';
+import stringToNumber from '../../src/stringToNumber.js';
 import testGrammar from '../testGrammar.js';
 
 const functions = require('./functions.json');
